MOR-410: buttons with customisable fields work in focus mode
